split state_machine/mocks into several files
