+++ rollup-interface/src/state_machine/mocks/mod.rs
@@ -1,0 +1,9 @@
+//! Defines mock instantiations of many important traits, which are useful
+//! for testing, fuzzing, and benchmarking.
+
+mod validity_condition;
+mod zk_vm;
+mod da;
+pub use validity_condition::TestValidityCond;
+pub use zk_vm::{MockZkvm, MockCodeCommitment, MockProof};
+pub use da::{TestBlob, TestBlock, TestBlockHeader, TestHash, MockAddress, MockDaSpec};
